ask binary rather than unconditional pkg manager call
 calling pkg manager induce delay and locking issue
 clean up

+++ dscripts/d01.nft.sh
@@ -4,9 +4,9 @@ lock_file=/tmp/$(basename $0 .sh).lock
 srcdir=$(cd $(dirname $0); pwd)
 # Exit the script on errors:
 set -e
-trap '{ echo "$0 FAILED on line $LINENO!; rm -f $0 ;}" | tee ${srcdir}/$(basename $0).log' ERR
+trap '{ echo "$0 FAILED on line $LINENO!; rm -f $0 ; }" | tee ${srcdir}/$(basename $0).log' ERR
 # clean up on exit
-trap "{ rm -f ${lock_file} ; rm -f $0; }" EXIT
+trap "{ rm -f ${lock_file} ; rm -f $0 ; }" EXIT
 
 # Catch unitialized variables:
 set -u
@@ -28,18 +28,37 @@ init() {
     pkg="python3 nftables"
     rule_file='"/etc/nftables/postgres_patroni"'
 
+    {
+        python3 --version > /dev/null 2>&1
+        nft --version > /dev/null 2>&1
+    } || {
+
+        case "${os_id}" in
+            'rhel' | 'centos' | 'fedora')
+                if [ "${os_major_version_id}" -lt 8 ]; then
+                    yum install -y $pkg
+                else
+                    dnf install -y $pkg
+                fi
+                nftables_conf="/etc/sysconfig/nftables.conf"
+                ;;
+            "debian" | "ubuntu")
+                apt-get install -y $pkg
+                test -d /etc/nftables/ || mkdir -p /etc/nftables/
+                nftables_conf="/etc/nftables.conf"
+                ;;
+            *)
+                echo "unsupported release vendor: ${os_id}" 1>&2
+                exit 1
+                ;;
+        esac
+    }
+
     case "${os_id}" in
         'rhel' | 'centos' | 'fedora')
-            if [ "${os_major_version_id}" -lt 8 ]; then
-                yum install -y $pkg
-            else
-                dnf install -y $pkg
-            fi
             nftables_conf="/etc/sysconfig/nftables.conf"
             ;;
         "debian" | "ubuntu")
-            apt-get install -y $pkg
-            test -d /etc/nftables/ || mkdir -p /etc/nftables/
             nftables_conf="/etc/nftables.conf"
             ;;
         *)
@@ -47,6 +66,7 @@ init() {
             exit 1
             ;;
     esac
+
     if [ ! -f /etc/nftables/postgres_patroni ]; then
         touch /etc/nftables/postgres_patroni || exit 1
     fi
@@ -68,22 +88,21 @@ EOF
 nftables_enable() {
     case "${os_id}" in
         'rhel' | 'centos' | 'fedora' | 'debian' | 'ubuntu')
-            if ! systemctl is-enabled nftables; then
-                systemctl enable --now nftables
-            fi
-            if nft -c 'flush ruleset; include "/etc/nftables/postgres_patroni.nft";'; then
-                if [ "$(md5sum /etc/nftables/postgres_patroni.nft | cut -d ' ' -f 1)" == "$(md5sum /etc/nftables/postgres_patroni | cut -d ' ' -f 1)" ]; then
-                    rm -f /etc/nftables/postgres_patroni.nft
+            systemctl is-enabled nftables || systemctl enable --now nftables
+            if [ -f /etc/nftables/postgres_patroni.nft ]; then
+                if nft -c 'flush ruleset; include "/etc/nftables/postgres_patroni.nft";'; then
+                    if [ "$(md5sum /etc/nftables/postgres_patroni.nft | cut -d ' ' -f 1)" == "$(md5sum /etc/nftables/postgres_patroni | cut -d ' ' -f 1)" ]; then
+                        rm -f /etc/nftables/postgres_patroni.nft
+                    else
+                        mv /etc/nftables/postgres_patroni.nft /etc/nftables/postgres_patroni
+                        systemctl reload nftables
+                    fi
                 else
-                    mv /etc/nftables/postgres_patroni.nft /etc/nftables/postgres_patroni
-                    systemctl reload nftables
+                    rm -f /etc/nftables/postgres_patroni.nft
+                    exit 1
                 fi
-            else
-                rm -f /etc/nftables/postgres_patroni.nft
-                exit 1
+                systemctl status nftables
             fi
-            systemctl status nftables
-            return $?
             ;;
         *)
             echo "unsupported release vendor: ${os_id}" 1>&2
@@ -106,7 +125,7 @@ disable_firewalld () {
 }
 
 {
-    flock -n 9 || exit 1
+    flock -w 10 9 || exit 1
 
     case "$1" in
         'init')
